+++ SOTFEdit/Model/Events/LanguageChangedEvent.cs
@@ -1,0 +1,5 @@
+namespace SOTFEdit.Model.Events;
+
+public class LanguageChangedEvent
+{
+}